+++ src/main/java/com/ufrest/dataobject/ResponseObject.java
@@ -1,24 +1,24 @@
-package com.ufrest;
+package com.ufrest.dataobject;
 
 import java.io.Serializable;
 
 /**
  * This represents the JSON object that ultimately gets returned to the caller
  * The response code is actually the response code from the call to UF, not this service
- * (for example UF can return a 500 error but this service will still return a 200)
- * The message is some arbitrary message that'll probably change
+ * (for example UF can return a 500 exception but this service will still return a 200)
  */
+@SuppressWarnings("unused")
 public class ResponseObject implements Serializable {
     private boolean success;
     private int responseCode;
     private Object data;
-    private String message;
+    private String targetURL;
 
-    public ResponseObject(boolean success, int responseCode, Object data, String message) {
+    public ResponseObject(boolean success, int responseCode, Object data, String targetURL) {
         this.success = success;
         this.responseCode = responseCode;
         this.data = data;
-        this.message = message;
+        this.targetURL = targetURL;
     }
 
     public boolean isSuccess() {
@@ -45,11 +45,11 @@ public class ResponseObject implements Serializable {
         this.data = data;
     }
 
-    public String getMessage() {
-        return message;
+    public String getTargetURL() {
+        return targetURL;
     }
 
-    public void setMessage(String message) {
-        this.message = message;
+    public void setTargetURL(String targetURL) {
+        this.targetURL = targetURL;
     }
 }
